fix MSVC warning

+++ cub/cub/detail/mdspan_utils.cuh
@@ -42,7 +42,7 @@ template <typename IndexType, size_t... Extents>
 [[nodiscard]] _CCCL_API constexpr ::cuda::std::make_unsigned_t<IndexType>
 size(const ::cuda::std::extents<IndexType, Extents...>& ext)
 {
-  return cub::detail::size_range(ext, 0, ext.rank());
+  return cub::detail::size_range(ext, 0, static_cast<int>(ext.rank()));
 }
 
 template <bool IsLayoutRight, int Position, typename IndexType, size_t... E>
